+++ apps/docs/src/app/core/component-docs/global-config/examples/dialog-global-config-example/dialog-global-config-example.module.ts
@@ -3,8 +3,7 @@ import { DIALOG_DEFAULT_CONFIG, DialogConfig } from '@fundamental-ngx/core';
 
 const DEFAULT_CONFIG: DialogConfig = {
     draggable: true,
-    escKeyCloseable: false,
-    backdropClickCloseable: false
+    escKeyCloseable: false
 };
 
 @NgModule({