+++ e2e/wdio/core/pages/fixed-card-layout.po.ts
@@ -23,4 +23,16 @@ export class FixedCardLayoutPo extends CoreBaseComponentPo {
         waitForElDisplayed(this.root);
         waitForPresent(this.pageHeader);
     }
+
+    getScreenshotFolder(): object {
+        return super.getScreenshotFolder(this.url);
+    }
+
+    saveExampleBaselineScreenshot(specName: string = 'fixed-card-layout'): void {
+        super.saveExampleBaselineScreenshot(specName, this.getScreenshotFolder());
+    }
+
+    compareWithBaseline(specName: string = 'fixed-card-layout'): any {
+        return super.compareWithBaseline(specName, this.getScreenshotFolder());
+    }
 }