+++ e2e/wdio/platform/pages/date-time-picker.po.ts
@@ -62,4 +62,16 @@ export class DateTimePicker extends BaseComponentPo {
         waitForElDisplayed(this.root);
         waitForPresent(this.datePickerInput);
     }
+
+    getScreenshotFolder(): object {
+        return super.getScreenshotFolder(this.url);
+    }
+
+    saveExampleBaselineScreenshot(specName: string = 'date-time-picker'): void {
+        super.saveExampleBaselineScreenshot(specName, this.getScreenshotFolder());
+    }
+
+    compareWithBaseline(specName: string = 'date-time-picker'): any {
+        return super.compareWithBaseline(specName, this.getScreenshotFolder());
+    }
 }