+++ e2e/wdio/platform/tests/date-time-picker.e2e-spec.ts
@@ -187,8 +187,8 @@ describe('Datetime picker suite', function() {
 
     xdescribe('Check visual regression', function() {
         it('should check examples visual regression', () => {
-            dateTimePickerPage.saveExampleBaselineScreenshot('date-time-picker');
-            expect(dateTimePickerPage.compareWithBaseline('date-time-picker')).toBeLessThan(1);
+            dateTimePickerPage.saveExampleBaselineScreenshot();
+            expect(dateTimePickerPage.compareWithBaseline()).toBeLessThan(1);
         });
     });