+++ CHANGELOG.md
@@ -2,6 +2,70 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.29.0-rc.10](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.9...v0.29.0-rc.10) (2021-03-10)
+
+## [0.29.0-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.8...v0.29.0-rc.9) (2021-03-10)
+
+## [0.29.0-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.7...v0.29.0-rc.8) (2021-03-10)
+
+
+### Bug Fixes
+
+* (Core|Platform) ng xi18n command running on angular/cli@10 fails ([#4889](https://github.com/SAP/fundamental-ngx/issues/4889)) ([11a961f](https://github.com/SAP/fundamental-ngx/commit/11a961ff26d36cf403d1d7f3132c87d1404f4be2))
+
+## [0.29.0-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.6...v0.29.0-rc.7) (2021-03-10)
+
+
+### Bug Fixes
+
+* (Core|Platform) select general issues ([#4852](https://github.com/SAP/fundamental-ngx/issues/4852)) ([d970cdc](https://github.com/SAP/fundamental-ngx/commit/d970cdc9a91a68c0f0709368e8bd6bd0c80e5401))
+
+## [0.29.0-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.5...v0.29.0-rc.6) (2021-03-10)
+
+## [0.29.0-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.4...v0.29.0-rc.5) (2021-03-09)
+
+## [0.29.0-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.3...v0.29.0-rc.4) (2021-03-09)
+
+
+### Bug Fixes
+
+* (Core) Dynamic page init content resize ([#4886](https://github.com/SAP/fundamental-ngx/issues/4886)) ([edce21f](https://github.com/SAP/fundamental-ngx/commit/edce21fdb3e22b8544393feaeeb13c1047aacaac))
+
+## [0.29.0-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.2...v0.29.0-rc.3) (2021-03-07)
+
+## [0.29.0-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.1...v0.29.0-rc.2) (2021-03-04)
+
+
+### Features
+
+* (Core) introduce the status Indicator component ([#4746](https://github.com/SAP/fundamental-ngx/issues/4746)) ([8f5bc1f](https://github.com/SAP/fundamental-ngx/commit/8f5bc1f77c9cab494d2fa7045e546386956da25a))
+
+## [0.29.0-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.0...v0.29.0-rc.1) (2021-03-04)
+
+## [0.29.0-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.28.2-rc.5...v0.29.0-rc.0) (2021-03-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* Changed dialog default behaviour closeOnBackdrop to false - Now it won't close when user clicks outside of dialog
+
+* fix: (Core) Change default dialog behaviour to not close on backdrop click
+
+* Add inputId to associate with label, add z-index to button in flexible column layout
+
+* Add examples with inputId and labels for pickers
+
+* BREAKING CHANGE: Portal issues - dialog default behaviour, inputIds for pickers  (#4839) ([b7c4d2e](https://github.com/SAP/fundamental-ngx/commit/b7c4d2e56cf29dc33b5ffc5f99670d2ac6823ce1)), closes [#4839](https://github.com/SAP/fundamental-ngx/issues/4839)
+
+### [0.28.2-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.28.2-rc.4...v0.28.2-rc.5) (2021-03-04)
+
+### [0.28.2-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.28.2-rc.3...v0.28.2-rc.4) (2021-03-03)
+
+
+### Bug Fixes
+
+* (Core) Remove small shellbar unnecessary styles ([#4799](https://github.com/SAP/fundamental-ngx/issues/4799)) ([fa84af0](https://github.com/SAP/fundamental-ngx/commit/fa84af084d65952885fda2d67dc8e32da8d67fb2))
+
 ### [0.28.2-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.28.2-rc.2...v0.28.2-rc.3) (2021-03-03)